Check in code
The layout is kinda crappy at the moment but it's a start.

Compile and invoke Test to see it in action. The constants might need
some tweaking for your machine as they were optimized for mine and might
prompt to run for too long for it to be worth it.