Implement Samplesort
This is a different kind of parallel sort intended to be more efficient,
though my benchmarks show that my implementation isn't :D

Probably because it must pass through the array at least twice to build
up the buckets which must be far from ideal, but this preserves some
contiguity of memory as a result. If we didn't care, we would just
allocate PROC_COUNT more arrays of the size equal to the input and just
merge them afterwards, but that still sounds like it'd involve either
discontiguous arrays or at least two iterations.

It might be interesting to study in more detail why good ol' parallel
merge sort works around 2x faster than this but oh well.